Changed floats to decimal

+++ RepairShop.Domain/Model/Entities/Quote.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public RepairOrder RepairOrder { get; set; } = null!;
-        public float PartTotal { get; set; }
-        public float LaborTotal { get; set; }
+        public decimal PartTotal { get; set; }
+        public decimal LaborTotal { get; set; }
         public DateTime ExpiryDate { get; set; }
     }
 }
